Adicionado os middlware de autenticação nas rotas

+++ resources/cardapios.py
@@ -1,9 +1,9 @@
 from flask_restful import Resource, marshal, reqparse
 from helpers.database import db
 from helpers.logger import logger
+from helpers.auth.token_verifier import token_verify
 
-from model.cardapio import Cardapio,cardapioFields
-from model.cardapio_preparacao import CardapioPreparacao
+from model.cardapio import Cardapio, cardapioTokenFields
 from model.mensagem import Message, msgFields
 from model.empresa import Empresa
 
@@ -12,12 +12,28 @@ parser = reqparse.RequestParser()
 parser.add_argument("empresa", type=dict, help="Empresa não informada", required=True)
 
 class Cardapios(Resource):
-  def get(self):
-    cardapios = Cardapio.query.all()
-    logger.info("Cardápios listados com sucesso")
-    return marshal(cardapios, cardapioFields), 200
+  @token_verify
+  def get(self, tipo, refreshToken):
+    if tipo != 'proprietario' and tipo != 'gestor':
+      logger.error("Usuario sem autorizacao para acessar o cardapio")
 
-  def post(self):
+      codigo = Message(1, "Usuario sem autorização suficiente!")
+      return marshal(codigo, msgFields), 403
+
+    cardapios = Cardapio.query.all()
+
+    data = {'cardapio': cardapios, 'token': refreshToken}
+
+    logger.info("Cardápios listados com sucesso")
+    return marshal(data, cardapioTokenFields), 200
+
+  @token_verify
+  def post(self, tipo, refreshToken):
+    if tipo != 'proprietario' and tipo != 'gestor':
+      logger.error("Usuario sem autorizacao para acessar o cardapio")
+
+      codigo = Message(1, "Usuario sem autorização suficiente!")
+      return marshal(codigo, msgFields), 403
     args = parser.parse_args()
     try:
 
@@ -35,8 +51,10 @@ class Cardapios(Resource):
       db.session.add(cardapio)
       db.session.commit()
 
+      data = {'cardapio': cardapio, 'token': refreshToken}
+
       logger.info(f"Cardápio de id: {cardapio.id} criado com sucesso")
-      return marshal(cardapio, cardapioFields), 201
+      return marshal(data, cardapioTokenFields), 201
     except:
       logger.error("Erro ao cadastrar o Cardápio")
 
@@ -44,7 +62,14 @@ class Cardapios(Resource):
       return marshal(codigo, msgFields), 400
 
 class CardapioId(Resource):
-  def get(self, id):
+  @token_verify
+  def get(self, tipo, refreshToken, id):
+    if tipo != 'proprietario' and tipo != 'gestor':
+      logger.error("Usuario sem autorizacao para acessar o cardapio")
+
+      codigo = Message(1, "Usuario sem autorização suficiente!")
+      return marshal(codigo, msgFields), 403
+
     cardapio = Cardapio.query.get(id)
 
     if cardapio is None:
@@ -53,24 +78,29 @@ class CardapioId(Resource):
       codigo = Message(1, f"Cardápio de id: {id} não encontrado")
       return marshal(codigo, msgFields), 404
 
-    logger.info(f"Cardápio de id: {id} listado com sucesso")
-    return marshal(cardapio, cardapioFields), 200
+    data = {'cardapio': cardapio, 'token': refreshToken}
 
-  def delete(self, id):
+    logger.info(f"Cardápio de id: {id} listado com sucesso")
+    return marshal(data, cardapioTokenFields), 200
+
+  @token_verify
+  def delete(self, tipo, refreshToken, id):
+    if tipo != 'proprietario' and tipo != 'gestor':
+      logger.error("Usuario sem autorizacao para acessar o cardapio")
+
+      codigo = Message(1, "Usuario sem autorização suficiente!")
+      return marshal(codigo, msgFields), 403
+
     cardapioBd = Cardapio.query.get(id)
-    cardapioPreparacaoBd = CardapioPreparacao.query.filter_by(cardapio_id=id)
 
     if cardapioBd is None:
       logger.error(f"Cardápio de id: {id} não encontrado")
 
       codigo = Message(1, f"Cardápio de id: {id} não encontrado")
       return marshal(codigo, msgFields), 404
-    
-    for i in cardapioPreparacaoBd:
-        db.session.delete(i)
 
     db.session.delete(cardapioBd)
     db.session.commit()
 
     logger.info(f"Cardápio de id: {id} deletado com sucesso")
-    return {}, 200
+    return {'token': refreshToken}, 200
